requirements.txt
tiktoken==0.5.2
uvicorn==0.25.0
fastapi==0.108.0
starlette==0.32.0.post1
httpx==0.26.0
python-dotenv